requirements.txt
streamlit 
seaborn
requests
google-generativeai
bs4
numpy 
pandas
yfinance
plotly
python.dotenv
pandas_datareader
streamlit_lottie
streamlit_option_menu